More warning fixes (conflicting func prototypes/defs, ...)

+++ include/qmckl_trexio.h
@@ -177,12 +177,12 @@ qmckl_get_ao_basis_nucleus_index_device(qmckl_context_device context,
 
 qmckl_exit_code
 qmckl_get_ao_basis_shell_ang_mom_device(qmckl_context_device context,
-										double *shell_ang_mom,
+										int32_t *shell_ang_mom,
 										int64_t shell_num);
 
 qmckl_exit_code
 qmckl_get_ao_basis_shell_factor_device(qmckl_context_device context,
-									   int32_t *shell_factor,
+									   double *shell_factor,
 									   int64_t shell_num);
 
 qmckl_exit_code
